+++ client/babel.config.js
@@ -1,19 +1,15 @@
-module.exports = function(api) {
+module.exports = function (api) {
     api.cache(true)
 
     return {
         presets: [
+            '@babel/preset-env',
             [
-                '@babel/preset-env',
+                '@babel/preset-react',
                 {
-                    useBuiltIns: 'entry',
-                    corejs: '3',
-                    targets: {
-                        node: 'current',
-                    },
+                    runtime: 'automatic',
                 },
             ],
-            '@babel/preset-react',
         ],
         plugins: [
             [