+++ client/src/constants/apiConstants.js
@@ -1,12 +1,6 @@
-export const APP_STATUS_REJECTED = 'NOT_APPROVED'
-export const APP_STATUS_PENDING = 'PENDING'
 export const APP_STATUS_APPROVED = 'APPROVED'
-
-export const appStatusToUI = {
-    NOT_APPROVED: 'Rejected',
-    PENDING: 'Pending',
-    APPROVED: 'Approved',
-}
+export const APP_STATUS_PENDING = 'PENDING'
+export const APP_STATUS_REJECTED = 'NOT_APPROVED'
 
 // manager roles as defined in Auth0
 export const roles = {