+++ client/src/utils/AuthService.js
@@ -1,7 +1,4 @@
-
 export default class AuthService {
-
-
     setProfile(profile) {
         localStorage.setItem('profile', JSON.stringify(profile))
     }