+++ server/src/server/init-server.js
@@ -7,7 +7,7 @@ const Vision = require('@hapi/vision')
 
 const HapiSwagger = require('hapi-swagger')
 const Pino = require('hapi-pino')
-
+const Schmervice = require('@hapipal/schmervice')
 const options = require('../options/index.js')
 
 const staticFrontendRoutes = require('../plugins/staticFrontendRoutes')
@@ -15,6 +15,7 @@ const apiRoutes = require('../plugins/apiRoutes')
 const errorMapper = require('../plugins/errorMapper')
 const queryFilter = require('../plugins/queryFilter')
 const pagination = require('../plugins/pagination')
+const { createEmailService } = require('../services/EmailService')
 
 exports.init = async (knex, config) => {
     debug('Starting server...')
@@ -86,6 +87,10 @@ exports.init = async (knex, config) => {
         })
     }
 
+    await server.register(Schmervice)
+
+    await server.registerService(createEmailService)
+
     await server.register({
         plugin: staticFrontendRoutes,
     })
@@ -103,6 +108,7 @@ exports.init = async (knex, config) => {
             options: {
                 knex,
                 auth: config.auth,
+                config,
             },
         },
         {